+++ R/rlandfire-package.R
@@ -3,5 +3,6 @@
 
 ## usethis namespace: start
 #' @importFrom utils browseURL
+#' @importFrom utils unzip
 ## usethis namespace: end
 NULL